pyproject.toml
[tool.poetry]
name = "rsync_files_from_digitalocean"
version = "0.1.0"
description = "rsync files from a given folder against certain digital ocean servers"
authors = ["Mark Grandi <markgrandi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-digitalocean = "^1.16.0"
arrow = "^1.1.1"
logging_tree = "^1.9"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]
wheel = "^0.37.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
